messenger/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.3.4
    # via django
django==3.2
    # via djangorestframework
djangorestframework==3.12.4
    # via -r requirements.in
pytz==2021.1
    # via django
sqlparse==0.4.1
    # via django
